fix: guard exact typedef keys against non-stdlib spelling collisions
_typedef_spelling_targets() checked every derived candidate spelling
(stdlib-stripped or namespace-suffix forms) against non-stdlib record
spellings before registering it, but registered the typedef's exact key
unconditionally. Combined with direct-clang's typedef-scope-loss (a
namespaced alias losing its namespace at snapshot-production time), an
exact key could collide with an unrelated non-stdlib record sharing the
same bare name, resolving a public function taking that record by value
through the wrong typedef target instead.

Apply the same collision guard to the exact-key registration.

Also splits tests/test_type_reachability.py (pushed over the
AI-readiness 2000-line hard cap by this round's additions) into a new
sibling tests/test_type_reachability_mangling.py holding the
mangled-name-derived reachability tests -- a pure file split, no test
behavior changed.
